+++ sdk/keyvault/keyvault-certificates/test/public/mergeAndImport.spec.ts
@@ -34,7 +34,11 @@ describe("Certificates client - merge and import certificates", () => {
     recorder = authentication.recorder;
     keyVaultUrl = authentication.keyVaultUrl;
     credential = authentication.credential;
-    secretClient = new SecretClient(keyVaultUrl, credential, recorder.configureClientOptions({}));
+    secretClient = new SecretClient(
+      keyVaultUrl,
+      credential,
+      recorder.configureClientOptions({ verifyChallengeResource: false })
+    );
   });
 
   afterEach(async function () {