+++ sdk/keyvault/keyvault-keys/src/keysModels.ts
@@ -28,6 +28,12 @@ export interface KeyClientOptions extends ExtendedCommonClientOptions {
    * The version of the KeyVault's service API to make calls against.
    */
   serviceVersion?: string;
+
+  /**
+   * Whether to verify the authentication challenge resource matches the Key Vault or Managed HSM domain.
+   * Defaults to true.
+   */
+  verifyChallengeResource?: boolean;
 }
 
 /**