+++ sdk/keyvault/keyvault-secrets/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 4.6.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+- Verify the challenge resource matches the vault domain.
+  This should affect few customers who can set `verifyChallengeResource` in the options bag to `false` to disable.
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 4.5.1 (2022-08-15)
 
 ### Bugs Fixed